Bug fix: URL bar was too trigger happy
It would change the URL in the middle of typing in the URL, making the app extremely unusable.

I debugged this a bit and it was seemingly caused by the double bindings. It seems like a somewhat bizarre way to construct the WPF bindings. Doing it using a single binding instead makes more sense, and the way to trigger the URL change is now the Enter key, or moving the focus away from the URL bar.

I've verified that navigating around still updates the URLs displayed in the URL bar.